Split out inner classes
