+++ benchmarks/src/main/java/com/fasterxml/jackson/module/blackbird/SomeEnum.java
@@ -1,0 +1,5 @@
+package com.fasterxml.jackson.module.blackbird;
+
+public enum SomeEnum {
+    EA, EB, EC, ED
+}
